Plot the IB curve
This script plots the IB curve by computing the IB point for the same
beta multiple times, and taking the "optimal" result.

It behaves as expected for the case of |X| = |Y|, but does not work as
expected when |X|=50 and |Y| = 10. Instead of getting a nice curve, we
get a "cloud" of points. This is likely a convergence error on the part
of the IB.m script, which is to be improved in a future commit.