Use JS-Divergence for convergence
The Kullback-Leibler divergence was giving negative values when P and Q
were sufficiently close (i.e. when the absolute value of their
difference is within a tolerance of 0.00000001). Thus, we now assume
that P = Q if max(abs(P - Q)) < 10^-8. If we somehow get a negative
value for Dkl, we set it to INF (since there must have been a division
by something so close to zero that the sum became negative).

We also use the JS-divergence for measuring the convergence of the IB
algorithm between iterations of Q(T|X).