Began implementing Generalized bottleneck
Started writing the function for the generalized bottleneck functional.
So far, this has implemented the updates for Q(t), Q(Y|T), and
initializes Q(T|X) randomly.

Next steps involve running iterations to converge this for a given beta.
This bottleneck may be the one to do it multiple times, or that can be
left for anything that calls this function (i.e. any other function that
calls this should run this N times to guarantee convergence to some more
optimal L value).

This also needs to have an updateQtgx function which chooses a method of
updating the functional depending on alpha, beta, and gamma.