Fixed issues that result in a negative value being placed in the distribution.
If for some reason the sum of the distribution along the given dimension
is 1 + eps, we get a negative value in the distribution (which doesn't
make sense). Thus, we swap the order of how we do things.

First, we re-normalize along the desired dimension.

Second, we set values below 10^-15 to zero.

Third, we throw away decimals below 10^-15 by taking
floor(P*10^15)/10^15.

Finally, we set the final value of the distribution to 1 - sum(other
values).

This appears to work for all test cases, even when sum(sum(P)) == 1 +
eps.