Starting to compute curve once betas are known
Added a waitbar to the beta finder.

Created a function which will get the curve points H(T), H_gamma(T),
I(X;T) and I(Y;T) for given betas. This is to be run after the betas
have been computed, so as to compute the positions on the curve for all
of these.

The curve will minimize the L functional for the generalized bottleneck
functional, but the function will return the positions on all planes for
all the inputted betas. These can then be plotted.