Finished the curve point computations
The algorithm now performs a similar task to the plotIB function, in
that it can compute the "optimal" L-value for the input points. It also
outputs the H(T), H_gamma(T) - H(T|X), I(X;T), and I(T;Y) values so they
can be plotted in the main function.