Handling the display cases for input flags
The input string for which item to display is now handled, where the
third plot is only displayed if the alpha and gamma values are not
chosen to show the (D)IB. It also automatically chooses which of the
IB/DIB plots to display.

It looks like the partition search function does not work as desired -
in some cases, the bottleneck function returns a sub-optimal Hga to
start and so the beta is much higher than it should be, which means
future betas cannot meet the HgaToFind range. This needs to be fixed
ASAP if the partition function is going to work as desired.