Finding better beta values
This update improves the partition search. Before, we just found any
beta which worked and stopped the search if the boundaries of the
bisection method were the same. Now, in that situation we know that we
must have skipped over an optimal beta. Thus, we choose to instead move
the boundaries around until we find an optimal beta value (with
restrictions on where those boundaries can move). This tends to find
reasonable beta values, especially if it's done a few times.

Also, if the curve results in a suboptimal Hga or I(T;Y) value (i.e. the
one for "this' beta is less than the one for the "previous" beta) the
loop is repeated, to try to find more globally optimal outputs. So far,
the plots appear to be forming nicer curves than before.