Update to plot all three figures
Can now plot the Generalized IB curve for gamma,alpha ~= 1.

For some values of alpha and gamma, we can have Hga negative (or
strictly positive) so we add a function to search for this minimum Hga
when we find a partition.

In the (D)IB case, when the previous Hga value is larger than the
current one we can start the search again for the current beta. This is
not done for general curve because it seems to freeze.

If it gets stuck a bunch of times for a fixed beta, we can interpolate
beta in between and try to reset. This is not ideal, and doesn't work as
well as we'd hope. It should eventually be removed in favour of other
fixes.