Replaced computation of points with optimalbottle()
We were computing the Hga value constantly in the loop, and running an
optimization which we will need to do in other places. Rather than
re-write the optimization loop, I just placed the optimization into its
own function (i.e. optimalbottle()).

Also, we increased the default max iterations for this optimization loop
from 50 to 100 because it works much better at avoiding loopback.