Changed loopback fix
The attempt to force the DIB to redo the bottleneck if Hga <
previousHga, or Iyt < previousIyt, does not work for very large matrices
(as those we get when we are doing clustering). Instead, the DIB only
tries to make sure the current L is less than the previousL (within a
tolerance epsilon).

This works okay for the standard DIB, but not great for the REnyi-DIB.
This needs to be improved more to get a good curve.