Finished the bottlecluster function
This function computes the bottleneck curve for a set of points using
the results from Strouse and Schwab's "Geometric Clustering" paper. This
can compute the clusters for hardcoded betas - once the DIB can search
for the beta partitions, this will work much better.

It also returns a struct containing all the desired data, such as the
DIB curve, the beta values, the points, etc... so that we can analyze
the data later