Added an analyzing function for the laplacian data.
Also changed the plots to show the correct legend when we do not
generate a DIB partition contour plot of q(c|x).

This commit is the final one to get all data required for the
end-of-summer report related to the clustering (though it is much later than expected). More
commits may come for plots of the Renyi-bottleneck.